Modified according to QA testing

+++ frontend/src/app/modules/employee/pages/assign-task/assign-task.component.ts
@@ -52,12 +52,17 @@ export class AssignTaskComponent implements OnInit {
         });
     }
 
-    updatePermission(employee: any, task: any): void {
+    async updatePermission(employee: any, task: any, module: any): Promise<any> {
+        if(this.isPermissionLoading(employee, task))
+            return;
         this.updatePermissionLoading(employee, task, true);
         if (this.hasPermission(employee, task)) {
             this.serviceAdapter.deletePermission(employee, task);
         } else {
-            this.serviceAdapter.addPermission(employee, task);
+            await this.serviceAdapter.addPermission(employee, task);
+            if(TASK_PERMISSION_LIST.find(taskPermission => taskPermission.modulePath == module.path && taskPermission.taskPath == task.path)) {
+                this.openInPagePermissionDialog(module, task, employee);
+            }
         }
     }
 
@@ -85,11 +90,7 @@ export class AssignTaskComponent implements OnInit {
         return employee.permissionLoading && task.permissionLoading;
     }
 
-    openInPagePermissionDialog(module, task, employee) {
-        if(this.isDisabled(module, task, employee)){
-            return;
-        }
-        
+    openInPagePermissionDialog(module, task, employee) {        
         const existingPermission = this.hasPermission(employee, task);
         const openedDialog = this.dialog.open(InPagePermissionDialogComponent, {
             data: {
@@ -113,7 +114,9 @@ export class AssignTaskComponent implements OnInit {
         });
     }
 
-    hasInPageTaskPermission(module, task): boolean {
+    hasInPageTaskPermission(module, task, employee): boolean {
+        if(this.isDisabled(module, task, employee))
+            return false;
         if (TASK_PERMISSION_LIST.find(taskPermission => taskPermission.modulePath == module.path && taskPermission.taskPath == task.path))
             return true;
         return false;
